// relive-webapp/src/app/interfaces/user-measurement.ts
export interface UserMeasurement {
    MeasureID: number;
    CustomerID : number;
    MeasureDate: Date;
    TotalWeight: number;
    Height: number;
    BMI: number;
    BMR: number;
    AbdominalFatPercentage: number;
    FatPercentage: number;
    Muscles: number; 
    Bones: number;
    Liquids: number;
    HipCircumference: number;
    HandCircumference: number;
    ThighCircumference: number;
    ChestCircumference: number
  }